Update interaction_plots.R
Added an intitial na.omit() function upon data importing. It selects the desired columns first so that, for example, if an unused phenotype column has NA values, those rows will not be filtered out and remove data points where the selected phenotype column does have data. I have also fixed the SNP_name variable which was incorrectly not referring back to the user-input for the SNP.